+++ packages/emitter/test/public-api-surface.test.ts
@@ -10,7 +10,6 @@ describe('Public API surface tests', () => {
       'EmitterWorkspace',
       'FileEmitter',
       'FileEmitterWorkspace',
-      'MultiEmitter',
     ]);
 
     expect(Exports.Emitter).to.be.a('function', 'class Emitter');