+++ packages/markdown-emitter/src/emitter/index.ts
@@ -1,13 +1,23 @@
-import { LinkedFormattedSourceFile, LinkedFormattedSymbol } from '@code-to-json/formatter-linker';
+import {
+  LinkedFormattedOutputData,
+  LinkedFormattedSourceFile,
+  LinkedFormattedSymbol,
+} from '@code-to-json/formatter-linker';
 import { isDefined } from '@code-to-json/utils';
 import { SysHost } from '@code-to-json/utils-ts';
-import { FileEmitter, FileEmitterWorkspace, EmitterState } from '@snap-doc/emitter';
+import { FileEmitter, FileEmitterWorkspace } from '@snap-doc/emitter';
 import * as debug from 'debug';
 import { heading, link, list, listItem, paragraph, rootWithTitle, text } from 'mdast-builder';
-import md from '../md';
+import * as remarkHtml from 'remark-html';
+import * as remarkParse from 'remark-parse';
+import * as remarkStringify from 'remark-stringify';
+import * as unified from 'unified';
 import { markdownForSourceFile } from '../md/file-generators/module';
 import { markdownForSymbolFile } from '../md/file-generators/symbol';
+import parserOptions from '../md/parser-options';
+import stringifyOptions from '../md/stringify-options';
 import MarkdownFileEmitterOptions from './options';
+import MarkdownEmitterState from './state';
 
 const log = debug('snap-doc:markdown-file-emitter');
 
@@ -15,11 +25,29 @@ export default class MarkdownFileEmitter extends FileEmitter<
   MarkdownFileEmitterOptions,
   FileEmitterWorkspace
 > {
+  protected md: unified.Processor = unified();
+
+  protected readonly ext: string = 'md';
+
   constructor(host: SysHost, options: MarkdownFileEmitterOptions) {
     super(host, options);
+    this.md.use(remarkParse, parserOptions);
+    if (options.html) {
+      this.md.use(remarkHtml);
+      this.ext = 'html';
+    } else {
+      this.md.use(remarkStringify, stringifyOptions);
+    }
   }
 
-  public async generate(state: EmitterState, workspace: FileEmitterWorkspace): Promise<void> {
+  protected async initializeState(data: LinkedFormattedOutputData): Promise<MarkdownEmitterState> {
+    return new MarkdownEmitterState(this.ext, data);
+  }
+
+  public async generate(
+    state: MarkdownEmitterState,
+    workspace: FileEmitterWorkspace,
+  ): Promise<void> {
     const {
       data: { sourceFiles: files },
     } = state;
@@ -55,7 +83,7 @@ if you want to replace existing content in the output directory`,
       .map(f => {
         // for each non-declaration file
         log(`Processing module: ${f.moduleName} (${f.path})`);
-        const outPath = this.pathInOutDir(workspace.pathFor(state, f));
+        const outPath = this.pathInOutDir(workspace.pathFor(state, f, this.ext));
         const content = this.contentForModule(state, workspace, f, {
           classes: classSymbols,
           types: typeSymbols,
@@ -66,7 +94,7 @@ if you want to replace existing content in the output directory`,
     // for each class collected along the way
     const classPaths = classSymbols.map(classSymbol => {
       // determine the output path
-      const outPath = this.pathInOutDir(workspace.pathFor(state, classSymbol));
+      const outPath = this.pathInOutDir(workspace.pathFor(state, classSymbol, this.ext));
 
       const content = this.contentForSymbol(state, workspace, classSymbol);
       // ...and write it
@@ -75,45 +103,40 @@ if you want to replace existing content in the output directory`,
     });
     // for each type collected along the way
     const typePaths = typeSymbols.map(typeSymbol => {
-      const outPath = this.pathInOutDir(workspace.pathFor(state, typeSymbol));
+      const outPath = this.pathInOutDir(workspace.pathFor(state, typeSymbol, this.ext));
       const content = this.contentForSymbol(state, workspace, typeSymbol);
       this.writeFileInOutDir(outPath, content);
       return [typeSymbol.text || typeSymbol.name, outPath];
     });
+
+    const tree = rootWithTitle(1, text(workspace.projectName), [
+      heading(2, text('Modules')),
+      paragraph(
+        list('unordered', modulePaths.map(mp => listItem(link(mp[1], mp[0], text(mp[0]))))),
+      ),
+      heading(2, text('Classes')),
+      paragraph(list('unordered', classPaths.map(cp => listItem(link(cp[1], cp[0], text(cp[0])))))),
+      heading(2, text('Types')),
+      paragraph(list('unordered', typePaths.map(tp => listItem(link(tp[1], tp[0], text(tp[0])))))),
+    ]);
+
     this.writeFileInOutDir(
-      'index.md',
-      md
-        .stringify(
-          rootWithTitle(1, text(workspace.projectName), [
-            heading(2, text('Modules')),
-            paragraph(
-              list('unordered', modulePaths.map(mp => listItem(link(mp[1], mp[0], text(mp[0]))))),
-            ),
-            heading(2, text('Classes')),
-            paragraph(
-              list('unordered', classPaths.map(cp => listItem(link(cp[1], cp[0], text(cp[0]))))),
-            ),
-            heading(2, text('Types')),
-            paragraph(
-              list('unordered', typePaths.map(tp => listItem(link(tp[1], tp[0], text(tp[0]))))),
-            ),
-          ]),
-        )
-        .trim(),
+      this.ext === 'html' ? 'index.html' : 'README.md',
+      this.md.stringify(tree).trim(),
     );
   }
 
   protected contentForSymbol(
-    state: EmitterState,
+    state: MarkdownEmitterState,
     workspace: FileEmitterWorkspace,
     sym: LinkedFormattedSymbol,
   ): string {
     const root = markdownForSymbolFile(state, workspace, sym);
-    return md.stringify(root).trim();
+    return this.md.stringify(root).trim();
   }
 
   protected contentForModule(
-    state: EmitterState,
+    state: MarkdownEmitterState,
     workspace: FileEmitterWorkspace,
     file: LinkedFormattedSourceFile,
     symbolsToSerialize: { classes: LinkedFormattedSymbol[]; types: LinkedFormattedSymbol[] },
@@ -139,6 +162,6 @@ if you want to replace existing content in the output directory`,
     log(`  └─ found types:\t${typeSymbols.map(s => s.text).join(', ')}`);
     symbolsToSerialize.classes.push(...classSymbols);
     symbolsToSerialize.types.push(...typeSymbols);
-    return md.stringify(sourceFileMarkdown).trim();
+    return this.md.stringify(sourceFileMarkdown).trim();
   }
 }