+++ packages/markdown-emitter/src/index.ts
@@ -1,2 +1,3 @@
 export { default as MarkdownFileEmitter } from './emitter';
+export { default as MarkdownFileEmitterState } from './emitter/state';
 export { default as MarkdownFileEmitterOptions } from './emitter/options';