+++ packages/markdown-emitter/types/remark-html.d.ts
@@ -1,0 +1,8 @@
+import { Attacher, Compiler } from 'unified';
+import { Node } from 'unist';
+
+interface RemarkHTML extends Attacher {
+  Compiler: Compiler;
+}
+declare const remarkHTML: RemarkHTML;
+export = remarkHTML;