+++ packages/markdown-emitter/types/to-vfile.d.ts
@@ -1,0 +1,3 @@
+declare const toVfile: any;
+
+export = toVfile;