+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,156 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.7...@snap-doc/utils@0.7.0) (2019-04-11)
+
+
+### Features
+
+* remove redundant remark-lint config ([68e966c](https://github.com/snap-doc/snap-doc/commit/68e966c))
+
+
+
+
+
+## [0.6.7](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.6...@snap-doc/utils@0.6.7) (2019-04-11)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.6.6](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.5...@snap-doc/utils@0.6.6) (2019-04-08)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.6.5](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.4...@snap-doc/utils@0.6.5) (2019-04-08)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.6.4](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.3...@snap-doc/utils@0.6.4) (2019-04-08)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.6.3](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.2...@snap-doc/utils@0.6.3) (2019-04-04)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.6.2](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.1...@snap-doc/utils@0.6.2) (2019-04-04)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.6.1](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.6.0...@snap-doc/utils@0.6.1) (2019-04-03)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+# [0.6.0](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.14...@snap-doc/utils@0.6.0) (2019-04-02)
+
+
+### Bug Fixes
+
+* add missing typescript package to utils ([371a11c](https://github.com/snap-doc/snap-doc/commit/371a11c))
+
+
+
+
+
+## [0.5.14](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.13...@snap-doc/utils@0.5.14) (2019-03-28)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.13](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.12...@snap-doc/utils@0.5.13) (2019-03-27)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.12](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.11...@snap-doc/utils@0.5.12) (2019-03-27)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.11](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.10...@snap-doc/utils@0.5.11) (2019-03-22)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.10](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.9...@snap-doc/utils@0.5.10) (2019-03-22)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.9](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.8...@snap-doc/utils@0.5.9) (2019-03-20)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.8](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.7...@snap-doc/utils@0.5.8) (2019-03-13)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.7](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.6...@snap-doc/utils@0.5.7) (2019-03-12)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
+## [0.5.6](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.5...@snap-doc/utils@0.5.6) (2019-03-11)
+
+**Note:** Version bump only for package @snap-doc/utils
+
+
+
+
+
 ## [0.5.5](https://github.com/snap-doc/snap-doc/compare/@snap-doc/utils@0.5.4...@snap-doc/utils@0.5.5) (2019-03-09)
 
 **Note:** Version bump only for package @snap-doc/utils